Crypto Trading Bot Python project/requirements.txt
python-binance
tabulate